Refactor to git-keeper-robot library.
To keep things organized, I started splitting up the functionality in
`VagrantControl.py`.  This led to the realization that this code should
be its own library.

To run the tests, you now must run `pip install -e .` in the
`git-keeper-robot` directory.
