+++ git-keeper-robot/gkeeprobot/keywords/ServerKeywords.py
@@ -1,0 +1,30 @@
+
+from gkeeprobot.control.ServerControl import ServerControl
+
+control = ServerControl()
+
+
+class ServerKeywords:
+
+    def start_gkeepd(self):
+        control.run('keeper', 'screen -S gkserver -d -m gkeepd')
+
+    def stop_gkeepd(self):
+        control.run('keeper', 'screen -S gkserver -X quit')
+
+    def expect_email(self, to_user, contains):
+        result = control.run_server_script('keeper', 'email_check.py', to_user, contains)
+        assert result == 'True'
+
+    def user_exists(self, username):
+        result = control.run_server_script('keeper', 'user_exists.py', username)
+        assert result == 'True'
+
+    def remove_user(self, username):
+        control.run('keeper', 'sudo userdel -r {} || :'.format(username))
+
+    def remove_file(self, username, filename):
+        control.run(username, 'rm {}'.format(filename))
+
+    def clear_email(self):
+        control.run('keeper', 'sudo find /email ! -name "mysmtpd.py" -type f -exec rm -f {} +')
